Implement vertex manipulation for relationship edges
- Add ability to add vertices to relationship edges via Shift+Click
- Add ability to delete vertices via right-click context menu
- Fix coordinate transformation issues in drag and drop
- Track drag position properly during relationship dragging
- Add visual vertex handles on selected relationships
- Update DiagramPainter to render vertex handles with blue circles
- Add onRelationshipVerticesChanged callback for immutable view updates
- Maintain mutable workspace copy in WorkspacePage for vertex updates
- Fix issue where clicking edges was repositioning nodes instead of adding vertices
- Add proper hit detection for vertices with 10-pixel radius
- Document vertex manipulation features in VERTEX_MANIPULATION.md

Also includes mobile platform enhancements and infrastructure improvements.

+++ lib/infrastructure/export/svg_exporter.dart
@@ -367,10 +367,10 @@ class SvgExporter implements DiagramExporter<String> {
           final y = (position.y is double)
               ? position.y
               : double.tryParse(position.y.toString()) ?? 0.0;
-          minX = min(minX as double, (x - 10.0) as double);
-          minY = min(minY as double, (y - 10.0) as double);
-          maxX = max(maxX as double, (x + width + 10.0) as double);
-          maxY = max(maxY as double, (y + height + 10.0) as double);
+          minX = min(minX, (x - 10.0) as double);
+          minY = min(minY, (y - 10.0) as double);
+          maxX = max(maxX, (x + width + 10.0) as double);
+          maxY = max(maxY, (y + height + 10.0) as double);
         }
       }
